Network access controlled at the interpreter level
Each individual interpreter can now be configured to use or not use the network with the Interpreter.NetworkingEnabled Boolean (as opposed to previously where it was a Boolean on the Roo module). Each network access attempt also raises a new event on the Interpreter called NetworkAccessed which is passed the URL and whether or not the Interpreter granted access or not.